+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Default Java icon in macOS Dock when using Java 9 or later [#655] (@WhiteBear60)
+- Issue with Instance settings allow an invalid Java path to be saved
 
 ### Misc